+++ domain/src/test/kotlin/fr/sacane/jmanager/domain/mock/Directory.kt
@@ -11,13 +11,13 @@ class Directory {
 
     companion object{
         val transactionInventories = mutableListOf(
-            Transaction(UUID.randomUUID(), "Piano", LocalDate.of(2022, Month.DECEMBER, 1), 450.toAmount(), true, tag = Tag("Fun")),
-            Transaction(UUID.randomUUID(), "Salary", LocalDate.now(), 3500.toAmount(), true, tag = Tag("Work")),
-            Transaction(UUID.randomUUID(), "SingLessons", LocalDate.now(), 450.toAmount(), true, tag = Tag("Fun")),
-            Transaction(UUID.randomUUID(), "Restaurant", LocalDate.of(2022, Month.DECEMBER, 4), 100.toAmount(), true, tag = Tag("Fun")),
-            Transaction(UUID.randomUUID(), "Laptop", LocalDate.of(2022, Month.DECEMBER, 31), 450.toAmount(), true, tag = Tag("Nothing")),
-            Transaction(UUID.randomUUID(), "", LocalDate.now(), 450.toAmount(), true, tag = Tag("Fun")),
-            Transaction(UUID.randomUUID(), "Money From testX", LocalDate.now(), 450.toAmount(), true, tag = Tag("Transaction"))
+            Transaction(UUID.randomUUID(), "Piano", LocalDate.of(2022, Month.DECEMBER, 1), 450.toAmount(), true, tag = Tag.Personal("Fun")),
+            Transaction(UUID.randomUUID(), "Salary", LocalDate.now(), 3500.toAmount(), true, tag = Tag.Personal("Work")),
+            Transaction(UUID.randomUUID(), "SingLessons", LocalDate.now(), 450.toAmount(), true, tag = Tag.Personal("Fun")),
+            Transaction(UUID.randomUUID(), "Restaurant", LocalDate.of(2022, Month.DECEMBER, 4), 100.toAmount(), true, tag = Tag.Personal("Fun")),
+            Transaction(UUID.randomUUID(), "Laptop", LocalDate.of(2022, Month.DECEMBER, 31), 450.toAmount(), true, tag = Tag.Personal("Nothing")),
+            Transaction(UUID.randomUUID(), "", LocalDate.now(), 450.toAmount(), true, tag = Tag.Personal("Fun")),
+            Transaction(UUID.randomUUID(), "Money From testX", LocalDate.now(), 450.toAmount(), true, tag = Tag.Personal("Transaction"))
         )
     }
 }